api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
google-cloud-bigquery==3.25.0
